fix: preserve runtime policy boundaries during credential admission restack
