+++ server/internal/agents/runtimepolicy/setup_test.go
@@ -1,0 +1,68 @@
+package runtimepolicy
+
+import (
+	"context"
+	"log"
+	"os"
+	"testing"
+
+	"github.com/jackc/pgx/v5/pgxpool"
+	accessrepo "github.com/speakeasy-api/gram/server/internal/access/repo"
+	"github.com/speakeasy-api/gram/server/internal/authz"
+	"github.com/speakeasy-api/gram/server/internal/conv"
+	orgrepo "github.com/speakeasy-api/gram/server/internal/organizations/repo"
+	"github.com/speakeasy-api/gram/server/internal/testenv"
+	"github.com/speakeasy-api/gram/server/internal/urn"
+	"github.com/stretchr/testify/require"
+)
+
+var cloneTestDatabase testenv.PostgresDBCloneFunc
+
+func TestMain(m *testing.M) {
+	ctx := context.Background()
+	container, clone, err := testenv.NewTestPostgres(ctx)
+	if err != nil {
+		log.Fatalf("launch test postgres: %v", err)
+	}
+	cloneTestDatabase = clone
+	code := m.Run()
+	if err := container.Terminate(ctx); err != nil {
+		log.Fatalf("terminate postgres container: %v", err)
+	}
+	os.Exit(code)
+}
+func newTestDB(t *testing.T) *pgxpool.Pool {
+	t.Helper()
+
+	conn, err := cloneTestDatabase(t, "testdb")
+	require.NoError(t, err)
+
+	return conn
+}
+
+func seedOrganization(t *testing.T, ctx context.Context, conn *pgxpool.Pool, organizationID string) {
+	t.Helper()
+
+	_, err := orgrepo.New(conn).UpsertOrganizationMetadata(ctx, orgrepo.UpsertOrganizationMetadataParams{
+		ID:       organizationID,
+		Name:     "Test Org",
+		Slug:     organizationID,
+		WorkosID: conv.PtrToPGText(conv.PtrEmpty("workos-org-" + organizationID)),
+	})
+	require.NoError(t, err)
+}
+
+func seedGrant(t *testing.T, ctx context.Context, conn *pgxpool.Pool, organizationID string, principal urn.Principal, scope authz.Scope, resource string) {
+	t.Helper()
+
+	selectors, err := authz.NewSelector(scope, resource).MarshalJSON()
+	require.NoError(t, err)
+
+	_, err = accessrepo.New(conn).UpsertPrincipalGrant(ctx, accessrepo.UpsertPrincipalGrantParams{
+		OrganizationID: organizationID,
+		PrincipalUrn:   principal,
+		Scope:          string(scope),
+		Selectors:      selectors,
+	})
+	require.NoError(t, err)
+}
